.history/requirements_20250908015820.txt
# requirements.txt
ccxt==4.1.45
pandas==2.1.3
TA-Lib==0.4.28
numpy==1.25.2
python-dotenv==1.0.0
schedule==1.2.0

---

# .env.example
# Bybit API credentials
BYBIT_API_KEY=your_api_key_here
BYBIT_API_SECRET=your_api_secret_here

# Trading configuration
TESTNET=true
SYMBOLS=BTC/USDT:USDT,ETH/USDT:USDT
LEVERAGE=3
RISK_PER_TRADE=0.01

# Strategy parameters
RSI_OVERSOLD=30
RSI_OVERBOUGHT=60
VOLUME_MULTIPLIER=1.8
TAKE_PROFIT=0.06
STOP_LOSS=0.03

# Risk management
MAX_CONSECUTIVE_LOSSES=3
PAUSE_HOURS=24
